Add files via upload
Use code with caution. Learn more

This code will first split the text into words. Then, it will iterate over the words and find the mood of each word. Finally, it will return the mood with the higher count.

To run this code, you can save it as a .cpp file and then compile it using a C++ compiler. Once the code is compiled, you can run it by typing the following command in a terminal:

./my_program

This will run the program and prompt you to enter some text. After you enter the text, the program will print the mood of the text.
profile picture
can we train this model?


Yes, you can train this model by providing it with a larger dataset of text and their corresponding moods. This will allow the model to learn the relationship between words and moods more accurately.

To train the model, you can use a supervised learning algorithm. This algorithm will learn the relationship between the words in the text and their corresponding moods by being trained on a dataset of labeled text. The labeled text will consist of pairs of text and moods, such as "I am happy" and "positive".

Once the model is trained, it can be used to find the mood of new text. The model will do this by predicting the mood of each word in the new text and then taking the majority vote of the predicted moods.

Here are some resources that you can use to train the model:

    A Gentle Introduction to Natural Language Processing: https://www.coursera.org/specializations/natural-language-processing
    Deep Learning with Python: https://www.manning.com/books/deep-learning-with-python
    Natural Language Processing with Python: https://www.nltk.org/book/